requirements.txt
preprocessing.p# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
timm==0.9.12  # PyTorch Image Models - for EfficientNet, ConvNeXt

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# Data Science
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Progress Bar
tqdm==4.66.1

# Configuration
PyYAML==6.0.1

# Utilities
python-dotenv==1.0.0

# Grad-CAM for PyTorch
pytorch-grad-cam==1.4.8

# Optional: For faster training with mixed precision
# No additional package needed - native in PyTorch